+++ packages/aws-amplify-react/src/Auth/Authenticator.jsx
@@ -40,7 +40,8 @@ export default class Authenticator extends Component {
         this.handleAuthEvent = this.handleAuthEvent.bind(this);
         this.onHubCapsule = this.onHubCapsule.bind(this);
 
-        this.state = { auth: props.authState || 'loading' };
+        this._initialAuthState = this.props.authState || 'signIn';
+        this.state = { auth: 'loading' };
         Hub.listen('auth', this);
     }
 
@@ -74,7 +75,7 @@ export default class Authenticator extends Component {
             })
             .catch(err => {
                 if (!this._isMounted) { return; }
-                Auth.signOut().then(() => this.handleStateChange('signedOut'));
+                Auth.signOut().then(() => this.handleStateChange(this._initialAuthState));
             });
     }