+++ packages/aws-amplify-react/src/Auth/FederatedSignIn.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Component } from 'react';
 
-import { JS, ConsoleLogger as Logger } from '@aws-amplify/core';
+import { JS, I18n, ConsoleLogger as Logger } from '@aws-amplify/core';
 import Auth from '@aws-amplify/auth';
 
 import AmplifyTheme from '../Amplify-UI/Amplify-UI-Theme';
@@ -97,7 +97,7 @@ export class FederatedButtons extends Component {
                 <div>
                 {this.OAuth(oauth_config)}
                 </div>
-                <Strike>or</Strike>
+                <Strike>{I18n.get('or')}</Strike>
             </div>
         )
     }