+++ packages/aws-amplify-react/src/Auth/Provider/withOAuth.jsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import { Component } from 'react';
 
-import { ConsoleLogger as Logger } from '@aws-amplify/core';
+import { I18n, ConsoleLogger as Logger } from '@aws-amplify/core';
 import Auth from '@aws-amplify/auth';
 import AmplifyTheme from '../../Amplify-UI/Amplify-UI-Theme';
+import { oAuthSignInButton } from '@aws-amplify/ui';
 import { 
     SignInButton, 
     SignInButtonContent
@@ -50,12 +51,13 @@ export default function withOAuth(Comp, options) {
 
 const Button = (props) => (
     <SignInButton
-        id="OAuth_signin_btn"
+        id={oAuthSignInButton}
         onClick={props.OAuthSignIn}
         theme={props.theme || AmplifyTheme}
+        variant={'oAuthSignInButton'}
     >
-        <SignInButtonContent>
-            {props.label || 'Sign in with AWS'}
+        <SignInButtonContent theme={props.theme || AmplifyTheme}>
+            {I18n.get(props.label || 'Sign in with AWS')}
         </SignInButtonContent>
     </SignInButton>
 )